### 2023-04-05 #5343 @editorialbot
[PRE REVIEW]: riccati: an adaptive, spectral solver for oscillatory ODEs
**Submitting author:** @fruzsinaagocs (<a href="http://orcid.org/0000-0002-1763-5884">Fruzsina Julia Agocs</a>)
**Repository:** https://github.com/fruzsinaagocs/riccati
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.13
