### 2023-04-20 #5343 @fruzsinaagocs
Hi @diehlpk, sure. From the JOSS table: mscroggs, krystophny, mseri, sixpearls, wingtorres, victorapm, mbarzegary (currently in a review).
