### 2023-05-01 #5343 @mseri
The issue still says Pre Review. I had a look at the paper, the documentation and the code. I installed the package, run the tests and tried a few examples. All seems in order according to the main guidelines.

The only issue I encountered is that for running the tests locally the suggested command is not workign for me
```
$ python -m pytest -v tests
platform darwin -- Python 3.11.3, pytest-7.3.1, pluggy-1.0.0 -- /Users/mseri/code/test/riccati/venv/bin/python3.11
cachedir: .pytest_cache
rootdir: /Users/mseri/code/test/riccati
collected 0 items

============================ no tests ran in 0.00s =============================
ERROR: file or directory not found: tests
```
and instead I needed to drop the last argument:
```
python3.11 -m pytest
platform darwin -- Python 3.11.3, pytest-7.3.1, pluggy-1.0.0
rootdir: /Users/mseri/code/test/riccati
collected 17 items

riccati/tests/test_riccati.py .................                          [100%]

=============================== warnings summary ===============================
riccati/chebutils.py:4
  /Users/mseri/code/test/riccati/riccati/chebutils.py:4: DeprecationWarning: invalid escape sequence '\s'
    """

riccati/chebutils.py:41
  /Users/mseri/code/test/riccati/riccati/chebutils.py:41: DeprecationWarning: invalid escape sequence '\s'
    """

riccati/chebutils.py:162
  /Users/mseri/code/test/riccati/riccati/chebutils.py:162: DeprecationWarning: invalid escape sequence '\c'
    """

riccati/stepsize.py:4
  /Users/mseri/code/test/riccati/riccati/stepsize.py:4: DeprecationWarning: invalid escape sequence '\p'
    """

riccati/stepsize.py:36
  /Users/mseri/code/test/riccati/riccati/stepsize.py:36: DeprecationWarning: invalid escape sequence '\m'
    """

riccati/step.py:72
  /Users/mseri/code/test/riccati/riccati/step.py:72: DeprecationWarning: invalid escape sequence '\p'
    """

riccati/solversetup.py:5
  /Users/mseri/code/test/riccati/riccati/solversetup.py:5: DeprecationWarning: invalid escape sequence '\o'
    """

riccati/tests/test_riccati.py::test_integration
  /Users/mseri/code/test/riccati/riccati/tests/test_riccati.py:23: RuntimeWarning: invalid value encountered in divide
    maxerr = max(np.abs((fs_est - fs)/fs))

riccati/tests/test_riccati.py::test_nonosc_step
riccati/tests/test_riccati.py::test_nonosc_evolve
riccati/tests/test_riccati.py::test_nonosc_evolve_backwards
  /Users/mseri/code/test/riccati/riccati/chebutils.py:289: FutureWarning: `rcond` parameter will change to the default of machine precision times ``max(M, N)`` where M and N are the input matrix dimensions.
  To use the future default and silence this warning we advise to pass `rcond=None`, to keep using the old, explicitly pass `rcond=-1`.
    y1, res, rank, sing = np.linalg.lstsq(D2ic, rhs) # NumPy solve only works for square matrices

riccati/tests/test_riccati.py::test_osc_evolve
riccati/tests/test_riccati.py::test_osc_evolve_backwards
  /Users/mseri/code/test/riccati/riccati/stepsize.py:91: RuntimeWarning: invalid value encountered in divide
    maxgerr = max(np.abs((gest - gana)/gana))

riccati/tests/test_riccati.py::test_osc_evolve
riccati/tests/test_riccati.py::test_osc_evolve_backwards
  /Users/mseri/code/test/riccati/riccati/evolve.py:76: RuntimeWarning: invalid value encountered in scalar divide
    hosc_ini = sign*min(1e8, np.abs(wnext/dwnext), np.abs(gnext/dgnext))

riccati/tests/test_riccati.py::test_nonosc_evolve
riccati/tests/test_riccati.py::test_nonosc_evolve_backwards
  /Users/mseri/code/test/riccati/riccati/stepsize.py:30: RuntimeWarning: divide by zero encountered in scalar divide
    if max(ws) > (1 + epsh)/abs(h):

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
======================= 17 passed, 17 warnings in 4.14s ========================
```

In the past there used to be an automated checklist to fill, is it still the case?
