### 2023-05-19 #5430 @mseri
Concerning the reproducibility and performance points: it is all in principle reproducible but I should write the code myself for the comparison plot on the right hand side of Figure 1.

It would be nice to add a jupyter notebook or a python script in a, e.g. , `benchmark/` folder, where the code to obtain the plot is provided. This plot appears also in the [arxiv preprint](https://arxiv.org/pdf/2212.06924.pdf) (see Figure 5, but I don't seem to find the code there either).
