### 2023-05-20 #5430 @fruzsinaagocs
@mseri Great suggestion. The code that generates Fig. 1. of the JOSS paper (and equivalent in the preprint) exists, I'll make it public shortly.
