### 2023-06-01 a6f329b
Added code to generate fig. 1. in JOSS paper.
