+++ joss-paper/fig1.py
@@ -1,0 +1,381 @@
+import numpy as np
+import riccati
+import scipy.special as sp
+from scipy.integrate import solve_ivp
+import matplotlib
+from matplotlib import pyplot as plt
+import math
+import mpmath
+import os
+from num2tex import num2tex
+from pathlib import Path
+import time
+import pyoscode
+from matplotlib.legend_handler import HandlerTuple
+import pandas
+import subprocess
+from matplotlib.ticker import LogLocator
+
+
+class HandlerTupleVertical(HandlerTuple):
+    def __init__(self, **kwargs):
+        HandlerTuple.__init__(self, **kwargs)
+
+    def create_artists(self, legend, orig_handle,
+                       xdescent, ydescent, width, height, fontsize, trans):
+        # How many lines are there.
+        numlines = len(orig_handle)
+        handler_map = legend.get_legend_handler_map()
+
+        # divide the vertical space where the lines will go
+        # into equal parts based on the number of lines
+        height_y = (height / numlines)
+
+        leglines = []
+        for i, handle in enumerate(orig_handle):
+            handler = legend.get_legend_handler(handler_map, handle)
+
+            legline = handler.create_artists(legend, handle,
+                                             xdescent,
+                                             (2*i + 1)*height_y,
+                                             width,
+                                             2*height,
+                                             fontsize, trans)
+            leglines.extend(legline)
+
+        return leglines
+
+def Bremer237(l, n, eps, epsh, outdir, rdc = True, wkbmarching = False,\
+              kummer = False, oscode = False, rk = False):
+    """
+    Solves problem (237) from Bremer's "On the numerical solution of second
+    order ordinary differential equations in the high-frequency regime" paper.
+    """
+    def w(x):
+        return l*np.sqrt(1 - x**2*np.cos(3*x))
+
+    def g(x):
+        return np.zeros_like(x)
+
+    # For the reference solution 
+    def f(t, y):
+        yp = np.zeros_like(y)
+        yp[0] = y[1]
+        yp[1] = -l**2*(1 - t**2*np.cos(3*t))*y[0]
+        return yp
+
+    xi = -1.0
+    xf = 1.0
+    eps = eps
+    yi = 0.0
+    dyi = l
+    yi_vec = np.array([yi, dyi])
+
+    # Utility function for rounding to n significant digits
+    round_to_n = lambda n, x: x if x == 0 else round(x, - int(math.floor(math.log10(abs(x)))) + (n-1))
+        
+    # Reference solution and its reported error
+    reftable = "./data/eq237.txt"
+    refarray = np.genfromtxt(reftable, delimiter=',')
+    ls = refarray[:,0]
+    ytrue = refarray[abs(ls -l) < 1e-8, 1]
+    errref = refarray[abs(ls -l) < 1e-8, 2]
+
+    if rdc:
+        print("riccati")
+        N = 1000 # Number of repetitions for timing
+        epsh = epsh
+        n = n
+        p = n
+        start = time.time_ns()
+        info = riccati.solversetup(w, g, n = n, p = p)
+        for i in range(N):
[…]
