+++ joss-paper/data/eq237.txt
@@ -1,0 +1,9 @@
+# Self-reported accuracy of Bremer's phase function method on Eq. (237) in Bremer 2018. "On the numerical solution of second order ordinary differential equations in the high-frequency regime".
+# Lambda, y(t = 1), relative error
+1e1,0.2913132934408612e0,7e-14 
+1e2,0.5294889561602804e0,5e-13 
+1e3,-0.6028749132401260e0,3e-12 
+1e4,-0.4813631690625038e0,5e-11 
+1e5,0.6558931145821987e0,3e-10 
+1e6,-0.4829009413372087e0,5e-9 
+1e7,-0.6634949630196019e0,4e-8 
