### 2023-06-01 9536c37
Moved runtime comparison plot in benchmarks/, added to docs, requirements needed for run added.
