+++ benchmarks/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+scipy
+matplotlib
+pandas
+pyoscode
