### 2023-06-01 #5430 @fruzsinaagocs
The script generating the figure in this paper should now be in `benchmarks/`, with instructions on how to run it in the [Examples ](https://riccati.readthedocs.io/en/latest/examples/#benchmarks) section of the docs.
