### 2023-06-02 #5430 @mseri
It almost work, since you use submodules in there, I woudl modify the section with:

```
[...]
The script to generate this figure is available in the `benchmarks/` folder in the Github repository.
First make sure that you have correctly initialized the git submodules in your cloned repository with

    git submodule init
    git submodule update

and that you have `gfortran` installed.
After [installing riccati from source](https://riccati.readthedocs.io/en/latest/installation/#from-source), and working from the installation directory [...]
```
