### 2023-06-07 #5430 @fruzsinaagocs
The changes made during the review are available in [v1.1.0.](https://github.com/fruzsinaagocs/riccati/releases/tag/v1.1.0)
Here's the Zenodo doi: https://doi.org/10.5281/zenodo.8015503
