+++ joss-paper/paper.bib
@@ -1,7 +1,7 @@
 @misc{ardc,
 author = {Agocs, F J and Barnett, A H},
 keywords = {Numerical Analysis (math.NA), FOS: Mathematics, FOS: Mathematics},
-title = {An adaptive spectral method for oscillatory second-order linear ODEs with frequency-independent cost},
+title = {An adaptive spectral method for oscillatory second-order linear {ODEs} with frequency-independent cost},
 publisher = {arXiv},
 year = {2022},
 doi = {10.48550/arxiv.2212.06924}
@@ -14,7 +14,7 @@ volume = {5},
 number = {56},
 pages = {2830}, 
 author = {Agocs, F J}, 
-title = {{(py)oscode: fast solutions of oscillatory ODEs}}, 
+title = {{(py)}oscode: fast solutions of oscillatory {ODEs}}, 
 journal = {Journal of Open Source Software},
 doi = {10.21105/joss.02830}
 }
@@ -63,7 +63,7 @@ doi={10.1017/S0962492900002750}
 }
 
 @article{wkbmarching1,
-title={{WKB-based scheme with adaptive step size control for the Schr{\"o}dinger equation in the highly oscillatory regime}},
+title={{WKB}-based scheme with adaptive step size control for the {S}chr{\"o}dinger equation in the highly oscillatory regime},
 author={K{\"o}rner, Jannis and Arnold, Anton and D{\"o}pfner, Kirian},
 journal={Journal of Computational and Applied Mathematics},
 volume={404},
@@ -74,7 +74,7 @@ doi = {10.1016/j.cam.2021.113905}
 }
 
 @article{wkbmarching2,
-title={{WKB-based schemes for the oscillatory 1D Schr{\"o}dinger equation in the semiclassical limit}},
+title={WKB-based schemes for the oscillatory 1{D} {S}chr{\"o}dinger equation in the semiclassical limit},
 author={Arnold, Anton and Abdallah, Naoufel Ben and Negulescu, Claudia},
 journal={SIAM journal on numerical analysis},
 volume={49},
@@ -99,8 +99,8 @@ author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
           Harris, Charles R. and Archibald, Anne M. and
           Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
           {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-          Computing in Python}},
+title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
+          Computing in Python},
 journal = {Nature Methods},
 year    = {2020},
 volume  = {17},
@@ -109,7 +109,7 @@ doi     = {10.1038/s41592-019-0686-2},
 }
 
 @article{dop853,
-title={A family of embedded Runge-Kutta formulae},
+title={A family of embedded {R}unge-{K}utta formulae},
 author={Dormand, John R and Prince, Peter J},
 journal={Journal of computational and applied mathematics},
 volume={6},
