### 2023-06-12 #5430 @danielskatz
I'm suggesting some minor changes in the paper and bib in https://github.com/fruzsinaagocs/riccati/pull/17 - please merge this, or let me know what you disagree with, then we can proceed.
